Added LTC back,

+++ src/script_pokemon_util_80F87D8.c
@@ -635,7 +635,7 @@ bool8 DoesPartyHaveEnigmaBerry(void)
     return hasItem;
 }
 
-void CreateScriptedWildMon(u16 species, u8 level, u16 item)
+void CreateScriptedWildMon(u16 species, u16 level, u16 item)
 {
     u8 heldItem[2];
